Move contents of GlobalDefines to Log
