+++ srcDLL/op2ext.cpp
@@ -3,7 +3,6 @@
 #include "StringConversion.h"
 #include "OP2Memory.h"
 #include "FileSystemHelper.h"
-#include "GlobalDefines.h"
 #include "Log.h"
 #include "WindowsModule.h"
 #define WIN32_LEAN_AND_MEAN
