+++ srcStatic/ConsoleArgumentParser.cpp
@@ -2,7 +2,7 @@
 #include "ConsoleArgumentParser.h"
 #include "StringConversion.h"
 #include "FileSystemHelper.h"
-#include "GlobalDefines.h"
+#include "Log.h"
 #include <stdexcept>
