+++ srcStatic/FileSystemHelper.cpp
@@ -1,5 +1,4 @@
 #include "FileSystemHelper.h"
-#include "GlobalDefines.h"
 #define WIN32_LEAN_AND_MEAN
 #include <windows.h>
 #include <cstddef>
