+++ srcStatic/Log.cpp
@@ -1,5 +1,8 @@
 #include "Log.h"
 #include "Logger.h"
+#include "FileSystemHelper.h"
+#define WIN32_LEAN_AND_MEAN
+#include <windows.h>
 
 
 // Use anonymous namespace to keep globals private
@@ -24,5 +27,33 @@ void Log(const std::string& message, const std::string& moduleName) {
 	if (logger) {
 		// Delegate to internal logger
 		logger->Log(message, moduleName);
+	}
+}
+
+
+
+bool modalDialogsDisabled = false;
+
+void OutputDebug(const std::string& message)
+{
+#ifdef DEBUG
+	OutputDebugString(message.c_str());
+#endif
+}
+
+void DisableModalDialogs()
+{
+	modalDialogsDisabled = true;
+}
+
+void PostErrorMessage(const std::string& errorMessage, const std::string& sourcePathFilename, long lineInSourceCode)
+{
+	// __FILE__ returns absolute filename. Strip the absolute path to reduce clutter in log output
+	auto sourceFilename = fs::path(sourcePathFilename).filename().string();
+
+	const std::string formattedMessage = sourceFilename + ", Line: " + std::to_string(lineInSourceCode) + ": " + errorMessage;
+	Log(formattedMessage);
+	if (!modalDialogsDisabled) {
+		MessageBoxA(nullptr, formattedMessage.c_str(), "Outpost 2 Error", MB_ICONERROR);
 	}
 }
