+++ srcStatic/OP2Memory.cpp
@@ -1,5 +1,5 @@
 #include "OP2Memory.h"
-#include "GlobalDefines.h"
+#include "Log.h"
 #define WIN32_LEAN_AND_MEAN
 #include <windows.h>
 #include <sstream>
