+++ srcStatic/VolList.cpp
@@ -1,6 +1,6 @@
 #include "VolList.h"
 #include "OP2Memory.h"
-#include "GlobalDefines.h"
+#include "Log.h"
 #include <utility>
